1. added CRUD rest apis for person and expense group 2. added split stragtegies 3. added data.sql for DB initialisation

+++ src/main/java/com/lld/splitwise/person/PersonRepository.java
@@ -1,8 +1,11 @@
 package com.lld.splitwise.person;
 
+import com.lld.splitwise.expenseGroup.ExpenseGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 @Repository
-public interface PersonRepository extends JpaRepository<Person,Integer> {
+public interface PersonRepository extends JpaRepository<Person,Long> {
 }
